+++ packages/nextra/src/locales.ts
@@ -38,7 +38,7 @@ export function locales(request: NextRequest) {
 
   const shouldHandleLocale =
     !/^\/(api|_next)\//.test(nextUrl.pathname) &&
-    !/\.(jpe?g|svg|png|webmanifest|xml|ico|txt)$/.test(nextUrl.pathname) &&
+    !/\.(jpe?g|svg|png|webmanifest|xml|ico|txt|mp4)$/.test(nextUrl.pathname) &&
     nextUrl.locale !== '' &&
     // not Server-Side Error page
     nextUrl.pathname !== '/500'