+++ examples/swr-site/CHANGELOG.md
@@ -4,6 +4,27 @@
 
 ### Patch Changes
 
+- Updated dependencies [71c1897]
+- Updated dependencies [51ec00f]
+- Updated dependencies [2e441b7]
+- Updated dependencies [d5aa17c]
+- Updated dependencies [7f697e9]
+- Updated dependencies [016828e]
+- Updated dependencies [da585a8]
+- Updated dependencies [863a750]
+- Updated dependencies [eae1993]
+- Updated dependencies [b3219c3]
+- Updated dependencies [ad8625c]
+- Updated dependencies [673a125]
+- Updated dependencies [26c7e20]
+- Updated dependencies [3e9e54f]
+  - nextra-theme-docs@2.2.15
+  - nextra@2.2.15
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [bcaba9c]
 - Updated dependencies [a683c84]
 - Updated dependencies [f300fe7]